refactor(Game, GameController): 게임이 진행 중인지 확인해 시도횟수만큼 반복해라
Game 에 시도 횟수를 기반으로 게임이 진행 중인지 판단하는 기능 구현:
- attemptCount 를 캡슐화
- 시도 횟수가 남아 있다면 true 를 반환
