+++ src/test/java/racingcar/model/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
